Datalake: Fix the failure of splitting and reading text files.
The reason is that when the size of the file to be read exceeds the range of int32, type overflow occurs, resulting in the reading failure. see issue https://code.hashdata.xyz/field-engineering/dev_collaboration/-/issues/404

+++ contrib/datalake_fdw/src/common/rewrLogical.cpp
@@ -235,7 +235,7 @@ void readLogical::extraFragmentLists(std::vector<ListContainer> &lists, List *fr
 		ListContainer obj;
 		List *fragment = (List*)lfirst(cell);
 		obj.keyName = strVal(list_nth(fragment, 0));
-		obj.size = atoi(strVal(list_nth(fragment, 1)));
+		obj.size = atoll(strVal(list_nth(fragment, 1)));
 		lists.push_back(obj);
 	}
 	sort(lists.begin(), lists.end(), cmp);
